requirements.txt
gepa
portkey-ai
python-dotenv